api_kafka/kafka/ui/requirements.txt
streamlit
typing_extensions
psycopg2-binary